fix: call checkOrientation in setup. Fixes #20.
Moved all setup work that is only done once in the program lifetime to arduino setup().
Reason was issue #20, by reversing the coil connection the stepper could go wild on activation if the
control threshold is exceeded. Because this USUALLY is not the case this bug has not been noticed.